third_party/WebKit/LayoutTests/fast/forms/textarea-resize-above-min-size-set-and-below-initial-size-expected.txt

Test for resizing the Textarea above minimum size set and below it's initial size.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById("textInputID").style.width is "250px"
PASS document.getElementById("textInputID").style.height; is "250px"
PASS document.getElementById("textInputID").style.width is "250px"
PASS document.getElementById("textInputID").style.height; is "250px"
PASS document.getElementById("textInputID").style.width is "250px"
PASS document.getElementById("textInputID").style.height; is "250px"
PASS successfullyParsed is true

TEST COMPLETE
